make this source v3
Had major singularity issues with re-running this after updating vast +- fishstatutils. luckily was able to work only on UW desktop running the entire v3 script from scratch with NO package updates, and not loading in any datatsets. As evidence in runs from 04-27 the estimates are slightly different, not sure why yet, but the figure shows they're in the ballpark of what was generated in January. For selftest this needs to be a LIVE session so the code is therefore forced.