update for testing on laptop (memory error)
I wanted to see what happened if I ran vast v12+ on my laptop, for fear of updating and breaking everything on desktop, but got a memory error. this prompted me to abandon this clunky script and make idxMaster.R, which sources the data, record, etc made in the first few hundred lines here.