build custom extrapolation grid & plot
based on guidance here: https://github.com/James-Thorson-NOAA/VAST/wiki/Creating-an-extrapolation-grid

this uses the subarea shapefile to make the xtrap grid. my hope is that this can be used to generate PREDICTIONS at the subarea scale. For now we aren't including any BC data so we will simply ignore whatever gets produced for B2, B3.