standalone script for Cal Current at strata
to minimize the number of sel pars, I wanted to include the Triennial and WCGBTS surveys into a single standardization, with predictions at each of the C1, C2 strata. These are pre-split at the 1995 time block and we can monkey with selex at that same break as well as 2003. This supercedes an older script which I developed early 2021 which did effectively the same thing but predicted estimates for the entire coast only.