correct strata, nonencounter=FALSE and fieldconfig
I set up the fieldconfig here to match what is done by RACE. Last week I tried a run on this with nonencounter=TRUE (and the wrong stratification) which literally took DAYS to run  (that is 2022-01-17-AK_500nonEncounter). I didn't end up using that one in the OM. Going to experiment with correct strata and corrected fieldconfig -- I'm not positive if it was the fieldconfig or the interpolation which made it take forever.